app can now receive subtitles for YouTube videos by URL
A new feature has been added to the app: the app can now receive subtitles for YouTube videos by URL. The resulting subtitle text is used as the context for a DeepPavlov-based chatbot. Now the chatbot is able to answer questions using the text of the subtitles as the basis for the answers.